feat: enable configuring with profile url
Instead of having to search for the user id, the script will now take
the full user profile url (example "https://gamedev.lgbt/@nokorpo") and
get the user id with a request.

This commit also adds more documentation for the node properties. And it
hides the now "redundant" data like domain, username and user id from
the Editor's interface. But they are kept as exported properties so they
are saved with the scene whenever they are edited by the script.